refactor: fixed typing for ClassroomEditForm
