+++ src/types/classroom.types.ts
@@ -1,0 +1,4 @@
+export interface ClassroomEditState {
+  name: string;
+  classroom_id: string;
+}
